pagraph/requirements.txt
attrs==19.3.0
autobahn==17.10.1
Automat==0.8.0
blinker==1.4
cbor==1.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cloud-init==21.2
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
Cython==0.29.14
dataclasses==0.6
dbus-python==1.2.16
dgl-cu102==0.4.3
distro==1.4.0
distro-info===0.23ubuntu1
entrypoints==0.3
future==0.18.2
gitdb==4.0.9
GitPython==3.1.27
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
Jinja2==2.10.1
joblib==1.1.0
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
line-profiler==3.5.1
littleutils==0.2.2
llvmlite==0.39.0
lz4==3.0.2+dfsg
MarkupSafe==1.1.0
more-itertools==4.2.0
netifaces==0.10.4
networkx==2.8.5
numba==0.56.0
numpy==1.18.5
oauthlib==3.1.0
ogb==1.0.1
outdated==0.2.1
pandas==1.4.3
pexpect==4.6.0
Pillow==9.2.0
py-ubjson==0.14.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
PyQRCode==1.2.1
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.5
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-snappy==0.5.3
PyTrie==0.2
pytz==2022.2.1
PyYAML==5.3.1
requests==2.22.0
requests-unixsocket==0.2.0
scikit-learn==1.1.2
scipy==1.9.0
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
smmap==5.0.0
sos==4.1
ssh-import-id==5.10
systemd-python==234
threadpoolctl==3.1.0
torch==1.8.1+cu102
torchaudio==0.8.1
torchvision==0.9.1+cu102
tqdm==4.64.0
Twisted==18.9.0
txaio==2.10.0
typing-extensions==4.3.0
u-msgpack-python==2.1
ubuntu-advantage-tools==27.2
ufw==0.36
urllib3==1.25.8
wadllib==1.3.3
websockify==0.9.0
wsaccel==0.6.2
zipp==1.0.0
zope.interface==4.7.1
